Убрана отсылка к наименованию класс + Impl
