+++ ru.javacourse.atm/src/main/java/businesslogic/atm/ATM.java
@@ -1,5 +1,18 @@
 package businesslogic.atm;
 
-public interface ATMWorker {
-    void doATM();
+import businesslogic.scenarioautomat.ScenarioAutomat;
+import hardware.adapter.HardwareDecoratorAdapter;
+import hardware.adapter.HardwareDecorator;
+
+public class ATM implements ATMWorker{
+    private ScenarioAutomat scenarioAutomat;
+
+    public void doATM() {
+
+        HardwareDecoratorAdapter hardwareAdapter = new HardwareDecorator();
+        scenarioAutomat = new ScenarioAutomat();
+        do {
+            scenarioAutomat.playScenarios(System.in, System.out, hardwareAdapter);
+        } while (hardwareAdapter.selfCheck());
+    }
 }
