+++ ru.javacourse.atm/src/main/java/businesslogic/atm/ATMWorker.java
@@ -1,5 +1,5 @@
 package businesslogic.atm;
 
 public interface ATMWorker {
-    boolean workATM();
+    void doATM();
 }
