+++ ru.javacourse.atm/src/main/java/businesslogic/scenarioautomat/ScenarioAutomat.java
@@ -1,7 +1,7 @@
 package businesslogic.scenarioautomat;
 
 import businesslogic.useraction.*;
-import hardware.adapter.HardwareAdapter;
+import hardware.adapter.HardwareDecoratorAdapter;
 import lombok.NonNull;
 import java.io.InputStream;
 import java.io.PrintStream;
@@ -10,7 +10,7 @@ import java.util.List;
 
 public class ScenarioAutomat {
     @NonNull
-    private final List<Action> actions;
+    private List<Action> actions;
 
     private Action getLastAction(){
         return actions.get(actions.size() - 1);
@@ -20,10 +20,10 @@ public class ScenarioAutomat {
         return actions.get(actions.size() - 1).executeAction();
     }
 
-    private void addSelfCheckActionIfNeed(@NonNull InputStream in, @NonNull PrintStream out, @NonNull HardwareAdapter hardwareAdapter) {
+    private void addSelfCheckActionIfNeed(@NonNull InputStream in, @NonNull PrintStream out, @NonNull HardwareDecoratorAdapter hardwareDecoratorAdapter) {
         if (actions.size() == 0) {
             actions.add(new Action(
-                    ActionTypes.SelfCheck, 1, new SelfCheck(), in, out, hardwareAdapter)
+                    ActionTypes.SelfCheck, 1, new SelfCheck(), in, out, hardwareDecoratorAdapter)
             );
         }
     }
@@ -47,7 +47,7 @@ public class ScenarioAutomat {
     }
 
 
-    public void playScenarios(@NonNull InputStream in, @NonNull PrintStream out, @NonNull hardwareDecoratorAdapter hardwareDecoratorAdapter) {
+    public void playScenarios(@NonNull InputStream in, @NonNull PrintStream out, @NonNull HardwareDecoratorAdapter hardwareDecoratorAdapter) {
         actions = new ArrayList<>();
         do {
             addSelfCheckActionIfNeed(in, out, hardwareDecoratorAdapter);
